ixl(4): Stop setting RS bit on every descriptor.
I thought I had seen an issue with TSO where traffic would stop passing
and the queues would get hung unless RS was set on every descriptor,
but this appears to have gone away.

Note that RS is only supposed to be set on the last descriptor of a packet/
TSO send. There must have been a bug elsewhere?

Signed-off-by: Eric Joyner <erj@freebsd.org>